requirements.txt
Flask==2.3.2
gunicorn==20.1.0
openai>=1.0.0
python-telegram-bot==20.7
requests==2.31.0